Resolve comments PR
