+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "test": "mocha ./test/NumbersValidator.test.js",
     "coverage": "c8 mocha ./test/NumbersValidator.test.js",
-    "lint": "eslint . --ext .js --fix",
+    "lint": "eslint . --ext .js --fix --format stylish ",
     "wdio": "wdio run wdio.conf.js"
   },
   "keywords": [
@@ -25,10 +25,8 @@
     "chai": "^4.3.10",
     "eslint": "^8.50.0",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-chai-friendly": "^0.7.2",
     "eslint-plugin-import": "^2.28.1",
-    "eslint-plugin-prettier": "^5.0.1",
     "eslint-plugin-wdio": "^8.8.7",
     "mocha": "^10.2.0",
     "mochawesome": "^7.1.3"
