+++ test/helpers/customClick.js
@@ -1,0 +1,7 @@
+async function customClick(element) {
+  await element.waitForDisplayed();
+  await element.waitForEnabled();
+  await element.click();
+}
+
+module.exports = { customClick };
