--- test/pageobjects/secure.page.js
@@ -1,16 +1,0 @@
-const { $ } = require('@wdio/globals');
-const Page = require('./page');
-
-/**
- * sub page containing specific selectors and methods for a specific page
- */
-class SecurePage extends Page {
-  /**
-   * define selectors using getter methods
-   */
-  get flashAlert() {
-    return $('#flash');
-  }
-}
-
-module.exports = new SecurePage();
