+++ test/task1/specs/DOMPage.test.js
@@ -1,0 +1,7 @@
+describe('Challenging DOM Page Tests', () => {
+  it('should have the correct title', async () => {
+    await browser.url('/challenging_dom');
+    const title = await $('h3=Challenging DOM');
+    await expect(title).toBeDisplayed();
+  });
+});
