+++ test/task1/specs/dynamicContent.test.js
@@ -1,0 +1,12 @@
+describe('Dynamic Content Page Tests', () => {
+  it('should load images correctly', async () => {
+    await browser.url('/dynamic_content');
+    const images = await $$('img');
+
+    const isDisplayedPromises = images.map((image, index) => image.isDisplayed().then(result => {
+      expect(result).toBe(true, `Image at index ${index} is not visible`);
+    }));
+
+    await Promise.all(isDisplayedPromises);
+  });
+});
